pyproject.toml
[tool.poetry]
name = "skeleton-django-sqlite-minmal"
version = "0.2.0"
description = "providing a basic django structure to start of your projects"
authors = ["oryon-dominik <oryon@cyberise.de>"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^3.1.7"
django-environ = "^0.4.5"
django-extensions = "^2.2.9"
django-widget-tweaks = "^1.4.8"
django-debug-toolbar = "^3.2"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
hide_code = "^0.5.5"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
